+++ main/src/main/java/org/hibernate/tool/ant/JDBCConfigurationTask.java
@@ -17,7 +17,7 @@ import org.hibernate.tool.api.metadata.MetadataDescriptor;
 import org.hibernate.tool.api.metadata.MetadataDescriptorFactory;
 import org.hibernate.tool.api.reveng.ReverseEngineeringSettings;
 import org.hibernate.tool.api.reveng.ReverseEngineeringStrategy;
-import org.hibernate.tool.util.ReflectHelper;
+import org.hibernate.tool.internal.util.ReflectHelper;
 
 
 /**
